% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssfcm.R
\name{calculate_distances}
\alias{calculate_distances}
\title{Calculating distances between two matrices}
\usage{
calculate_distances(X, V)
}
\arguments{
\item{X}{a matrix of dimension (N, p)}

\item{V}{a matrix of dimension (C, p)}
}
\value{
a matrix of dimension (N, C)
}
\description{
A base R approach to calculate distances between
rows of two matrices A and B that are of the same ncol.
}
\details{
TODO it should be moved to tests to compare with other, externally
imported methods because this function turned out to be very slow.
}
