% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssfcm.R
\name{dheve}
\alias{dheve}
\title{Creates DHE (stands for "distances horizontally exploded") and DVE
(stands for "distances vertically exploded") matrices.}
\usage{
dheve(A, vertical)
}
\arguments{
\item{A}{Matrix of size N x c.}

\item{vertical}{Boolean switch.
If `TRUE`, create DVE (vertical explosion).
If `FALSE`, create DHE (horizontal explosion).}
}
\value{
Matrix of size Nc x c
}
\description{
Creates DHE (stands for "distances horizontally exploded") and DVE
(stands for "distances vertically exploded") matrices.
}
\examples{
A <- matrix(c(1, 2, 3, 4, 5, 6), ncol = 2, byrow = TRUE)
AVE <- dheve(A, vertical = TRUE)
AHE <- dheve(A, vertical = FALSE)

}
