% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sspcm.R
\name{estimate_T}
\alias{estimate_T}
\title{Estimated T matrix with typicalities.}
\usage{
estimate_T(X, V, superF, alpha, function_dist, gammas)
}
\arguments{
\item{X}{a matrix *X* of dimension (N, p) containing predictor variables.}

\item{V}{a prototypes matrix of dimension (c, p)}

\item{superF}{the supervisionbinary matrix of the same dimension as *U*.}

\item{alpha}{a *N*-vector of observation-specific scaling factor values}

\item{function_dist}{A function of two arguments: matrices X and V of the same
number of columns.
It should return a matrix of (nrow(X) x nrow(V)) of distances
between each row of X and all rows of V.
In case of Euclidean distance, the result should not be squared!}

\item{gammas}{a *c*-vector of cluster-specific gamma parameter}
}
\description{
Estimated T matrix with typicalities.
}
