% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssfcm.R, R/sspcm.R
\name{estimate_V}
\alias{estimate_V}
\title{Equation to calculate clusters' prototypes matrix $\hat{V}$.}
\usage{
estimate_V(Phi, X)

estimate_V(Phi, X)
}
\arguments{
\item{Phi}{Matrix with weights of size N x c.}

\item{X}{Matrix with predictors of size N x p.}
}
\value{
Clusters' prototypes matrix of size c x p.

Clusters' prototypes matrix of size c x p.
}
\description{
Equation to calculate clusters' prototypes matrix $\hat{V}$.

Equation to calculate clusters' prototypes matrix $\hat{V}$.
}
