% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssfcm.R
\name{explode_dimension}
\alias{explode_dimension}
\title{Explode dimension of a 2d matrix to 3d array}
\usage{
explode_dimension(A, byrow = FALSE)
}
\arguments{
\item{A}{a \code{matrix}.}

\item{byrow}{states if the rows of the matrix should be replicated
\emph{vertically} (\code{byrow=FALSE})
or \emph{horizontally} (\code{byrow=TRUE}).
Note the default behaviour is to replicate \emph{horizontally}.

\emph{vertical} replication means each \emph{Ar} looks like

\if{html}{\out{<div class="sourceCode">}}\preformatted{A_r = [[a_r1, ..., a_r1],
       [... , ..., ... ],
       [a_rc, ..., a_rc]],
}\if{html}{\out{</div>}}

while \emph{horizontal} replication means each \emph{Ar} looks like

\if{html}{\out{<div class="sourceCode">}}\preformatted{A_r = [[a_r1, ..., a_rc],
       [... , ..., ... ],
       [a_r1, ..., a_rc]].
}\if{html}{\out{</div>}}}
}
\value{
a 3d \code{array} with appropriately replicated rows.
}
\description{
Turns a 2d matrix \emph{A} of size \emph{s} x \emph{c} into a 3d array by replicating
rows of the matrix \emph{A} either vertically or horizontally.

This function takes a matrix \emph{A}

\if{html}{\out{<div class="sourceCode">}}\preformatted{A = [[a_11, ..., a_1c],
     [... , ..., ... ]
     [a_s1, ..., a_sc]]
}\if{html}{\out{</div>}}

transforms each row \code{r} to form a separate 2d matrix \emph{Ar}
based on the \code{byrow} parameters, and returns a 3d array
\verb{[A1, ..., As]}.
}
\examples{
A <- matrix(c(1, 2, 3, 4), ncol=2)
# > A
#      [,1] [,2]
# [1,]    1    3
# [2,]    2    4
B <- explode_dimension(A)
# > B
# , , 1
#
#      [,1] [,2]
# [1,]    1    1
# [2,]    3    3
#
# , , 2
#
#      [,1] [,2]
# [1,]    2    2
# [2,]    4    4
}
