% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssfcm.R
\name{gamma}
\alias{gamma}
\title{Aggregates elements of DHE and DVE matrices in a step to build
evidence matrix E.
See vignette \code{vectorized-equations} for details.}
\usage{
gamma(dhe, dve)
}
\arguments{
\item{dhe}{DHE matrix of size Nc x c.}

\item{dve}{DVE matrix of size Nc x c.}
}
\value{
Matrix of size Nc x 1.
}
\description{
Aggregates elements of DHE and DVE matrices in a step to build
evidence matrix E.
See vignette \code{vectorized-equations} for details.
}
