% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssfcm.R
\name{TFCM}
\alias{TFCM}
\title{Time Fuzzy C-Means}
\usage{
TFCM(
  X,
  C,
  U = NULL,
  max_iter = 200,
  conv_criterion = 1e-04,
  function_dist = rdist::cdist,
  alpha = NULL
)
}
\description{
Time Fuzzy C-Means
}
