% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssfcm.R
\name{update_memberships}
\alias{update_memberships}
\title{Estimated U matrix with memberships}
\usage{
update_memberships(X, V, F_, alpha, fun.distances)
}
\arguments{
\item{X}{a matrix \emph{X} of dimension (N, p) containing predictor variables.}

\item{V}{a prototypes matrix of dimension (C, p)}

\item{F_}{the supervision  binary matrix of the same dimension as \emph{U}.}

\item{alpha}{the scaling factor, a floating point > 0.}

\item{fun.distances}{A function of two arguments: matrices X and V of the same
number of columns.
It should return a matrix of (nrow(X) x nrow(V)) of distances
between each row of X and all rows of V.
In case of Euclidean distance, the result should not be squared!}
}
\description{
The trick is to use 3d arrays for efficient calculations.
Let's take a D^2 matrix

\if{html}{\out{<div class="sourceCode">}}\preformatted{D62 = [[5, 145],
       [25, 85],
       [61, 41]]
}\if{html}{\out{</div>}}
}
\details{
\code{D_nominator} will be a 3d array of dimensions (2, 2, 3):
it contains 3 matrices 2x2.
Let's take a single 2x2 matrix from the array \code{D_nominator[,,1]}

\if{html}{\out{<div class="sourceCode">}}\preformatted{     [,1] [,2]
[1,]    5    5
[2,]  145  145
}\if{html}{\out{</div>}}

It is effectively

\if{html}{\out{<div class="sourceCode">}}\preformatted{          [,1]      [,2]
[1,]    d^2_11    d^2_11
[2,]    d^2_12    d^2_12
}\if{html}{\out{</div>}}

Let's now take \code{D_denominator[,,1]}

\if{html}{\out{<div class="sourceCode">}}\preformatted{     [,1] [,2]
[1,]    5  145
[2,]    5  145
}\if{html}{\out{</div>}}

It is effectively

\if{html}{\out{<div class="sourceCode">}}\preformatted{          [,1]      [,2]
[1,]    d^2_11    d^2_12
[2,]    d^2_11    d^2_12
}\if{html}{\out{</div>}}

If we now divide \code{D_nominator[,,1]} by \code{D_denominator[,,1]}, we get

\if{html}{\out{<div class="sourceCode">}}\preformatted{         [,1]     [,2]
[1,]    5/5    5/145
[2,]    145/5  145/145
}\if{html}{\out{</div>}}

which is
\link{,1}             \link{,2}
\link{1,}    d^2_11/d^2_11    d^2_11/d^2_12
\link{2,}    d^2_12/d^2_11    d^2_12/d^2_12
The \code{rowSums} of the above will give us a vector

\if{html}{\out{<div class="sourceCode">}}\preformatted{[1] e_11 e_12
}\if{html}{\out{</div>}}

\emph{almost} the first row of evidence matrix, which is the inverse of the above
i.e. a vector \verb{[1] 1/e_11 1/e_12}.

The vectorized operations on entire \code{D_nominator} and \code{D_denominator}
follow the above logic.
}
