% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssfcm.R
\name{xi_fcm}
\alias{xi_fcm}
\title{Rearranges elements of input matrix from a block matrix with vertical blocks
(column vectors) to a block matrix with horizontal blocks (row vectors).}
\usage{
xi_fcm(A, c)
}
\arguments{
\item{A}{Matrix of size Nc x 1.}

\item{c}{Number of columns in the wanted matrix.
Associated with the number of clusters.}
}
\value{
Matrix of size N x c.
}
\description{
Rearranges elements of input matrix from a block matrix with vertical blocks
(column vectors) to a block matrix with horizontal blocks (row vectors).
}
